+++ mediation/AdColony/source/android-library/adcolony/src/main/java/com/google/unity/mediation/adcolony/AdColonyUnityExtrasBuilder.java
@@ -14,16 +14,6 @@ import java.util.HashMap;
 public class AdColonyUnityExtrasBuilder implements AdNetworkExtras {
 
     /**
-     * Key to obtain Zone ID.
-     */
-    private static final String KEY_ZONE_ID = "zone_id";
-
-    /**
-     * Key to obtain User ID.
-     */
-    private static final String KEY_USER_ID = "user_id";
-
-    /**
      * Key to obtain the "show pre popup" option.
      */
     private static final String KEY_SHOW_PRE_POPUP = "show_pre_popup";
@@ -35,16 +25,6 @@ public class AdColonyUnityExtrasBuilder implements AdNetworkExtras {
 
     @Override
     public Bundle buildExtras(HashMap<String, String> extras) {
-        String zoneId = extras.get(KEY_ZONE_ID);
-        if (zoneId != null) {
-            AdColonyBundleBuilder.setZoneId(zoneId);
-        }
-
-        String userId = extras.get(KEY_USER_ID);
-        if (userId != null) {
-            AdColonyBundleBuilder.setUserId(userId);
-        }
-
         String showPrePopup = extras.get(KEY_SHOW_PRE_POPUP);
         if (showPrePopup != null) {
             AdColonyBundleBuilder.setShowPostPopup(Boolean.valueOf(showPrePopup));