+++ mediation/AdColony/source/plugin/Assets/GoogleMobileAds/Api/Mediation/AdColony/AdColonyMediationExtras.cs
@@ -4,11 +4,8 @@ namespace GoogleMobileAds.Api.Mediation.AdColony
 {
     public class AdColonyMediationExtras : MediationExtras
     {
-        public const string ZoneIdKey = "zone_id";
-        public const string UserIdKey = "user_id";
         public const string ShowPrePopupKey = "show_pre_popup";
         public const string ShowPostPopupKey = "show_post_popup";
-        public const string TestModeKey = "test_mode";
 
         public AdColonyMediationExtras() : base() { }
 
@@ -22,16 +19,6 @@ namespace GoogleMobileAds.Api.Mediation.AdColony
             get { return "AdColonyExtrasBuilder"; }
         }
 
-        public void SetZoneId(string zoneId)
-        {
-            this.Extras.Add(ZoneIdKey, zoneId);
-        }
-
-        public void SetUserId(string userId)
-        {
-            this.Extras.Add(UserIdKey, userId);
-        }
-
         public void SetShowPrePopup(bool showPrePopup)
         {
             this.Extras.Add(ShowPrePopupKey, showPrePopup.ToString());
@@ -40,11 +27,6 @@ namespace GoogleMobileAds.Api.Mediation.AdColony
         public void SetShowPostPopup(bool showPostPopup)
         {
             this.Extras.Add(ShowPostPopupKey, showPostPopup.ToString());
-        }
-
-        public void SetTestMode(bool testMode)
-        {
-            this.Extras.Add(TestModeKey, testMode.ToString());
         }
     }
 }